requirements.txt
azure-identity>=1.18.0
requests>=2.32.3
PyJWT>=2.9.0